requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LLM & Embeddings
openai==1.10.0
sentence-transformers==2.3.1
tiktoken==0.5.2

# Vector Store
faiss-cpu==1.7.4

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
ruff==0.1.13
mypy==1.8.0
